move log level to main panel
