+++ packages/studio/src/components/RenderModal/RenderModalBasic.tsx
@@ -1,11 +1,13 @@
-import type {Codec, ProResProfile} from '@remotion/renderer';
+import type {Codec, LogLevel, ProResProfile} from '@remotion/renderer';
 import {BrowserSafeApis} from '@remotion/renderer/client';
 import {NoReactAPIs} from '@remotion/renderer/pure';
+import type {ChangeEvent} from 'react';
 import React, {useCallback, useMemo} from 'react';
 import type {VideoConfig} from 'remotion';
 import {labelProResProfile} from '../../helpers/prores-labels';
 import {useFileExistence} from '../../helpers/use-file-existence';
 import {Checkmark} from '../../icons/Checkmark';
+import {Checkbox} from '../Checkbox';
 import type {ComboboxValue} from '../NewComposition/ComboBox';
 import {Combobox} from '../NewComposition/ComboBox';
 import {InputDragger} from '../NewComposition/InputDragger';
@@ -43,6 +45,8 @@ export const RenderModalBasic: React.FC<{
 	readonly endFrame: number;
 	readonly setStartFrame: React.Dispatch<React.SetStateAction<number | null>>;
 	readonly validationMessage: string | null;
+	readonly setVerboseLogging: React.Dispatch<React.SetStateAction<LogLevel>>;
+	readonly logLevel: LogLevel;
 }> = ({
 	renderMode,
 	imageFormatOptions,
@@ -60,6 +64,8 @@ export const RenderModalBasic: React.FC<{
 	setStartFrame,
 	startFrame,
 	validationMessage,
+	setVerboseLogging,
+	logLevel,
 }) => {
 	const existence = useFileExistence(outName);
 	const videoCodecOptions = useMemo((): ComboboxValue[] => {
@@ -127,6 +133,13 @@ export const RenderModalBasic: React.FC<{
 			setOutName(e.target.value);
 		},
 		[setOutName],
+	);
+
+	const onVerboseLoggingChanged = useCallback(
+		(e: ChangeEvent<HTMLInputElement>) => {
+			setVerboseLogging(e.target.checked ? 'verbose' : 'info');
+		},
+		[setVerboseLogging],
 	);
 
 	return (
@@ -205,6 +218,19 @@ export const RenderModalBasic: React.FC<{
 				validationMessage={validationMessage}
 				label={renderMode === 'sequence' ? 'Folder name' : 'Output name'}
 			/>
+			<div style={optionRow}>
+				<div style={label}>
+					Verbose logging <Spacing x={0.5} />
+					<OptionExplainerBubble id="logLevelOption" />
+				</div>
+				<div style={rightRow}>
+					<Checkbox
+						checked={logLevel === 'verbose'}
+						onChange={onVerboseLoggingChanged}
+						name="verbose-logging"
+					/>
+				</div>
+			</div>
 		</div>
 	);
 };
